Updated - Move menus into components layout folder
