+++ packages/admin/src/Views/Components/Layout/PageMenu.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace GetCandy\Hub\Views\Components;
+namespace GetCandy\Hub\Views\Components\Layout;
 
 use Illuminate\View\Component;
 
-class SideMenuLayout extends Component
+class PageMenu extends Component
 {
     public function __construct()
     {
@@ -17,6 +17,6 @@ class SideMenuLayout extends Component
      */
     public function render()
     {
-        return view('adminhub::components.side-menu-layout');
+        return view('adminhub::components.layout.page-menu');
     }
 }
